remove timeout from lock, used in video and stream player
Some operations might take a long time, like video prepare for example, and there's no way of knowing, what is a sufficient timeout. It's better just having it with no timeout, because all the actions are finite anyway and there's no risk of a deadlock.

If there was in fact a deadlock, it's better that it occurs so we can find it and eliminate the bug that caused it. Using timeout is actually a sub-optimal workaround.