track pids of servlets and make sure they're stopped between server restarts
this has the most impact when the server unexpectedly crashes. In that case the servlets keep running independently and if the server is started again at that point, it will duplicate the servlet processes. To prevent that, each servlet's pid is tracked in a pidfile and when server service is stopped, it explicitly attempts to stop each servlet just in case.