pause client aliveness check when there's a pending task
Some tasks may take a very long time and can take up so much cpu that entire network communication gets throttled which would cause client to not be able to notify the server that it's alive which would cause the server to reset the client after timeout. To prevent that the whole aliveness check will be paused whenever a task is being performed.