refactor system settings and screen orientation handling
Orientation handling got moved out of ScreenOrientationManager in the new version of front display, so we can make it independent and handle it any way we want. The best way is to make it a part of system settings.

It's also convenient to have system settings available both in client and server, so that's why I created FrontSystemSettings, which is just a proxy for server-side FSSystemSettings.