optimize shutdown
localmount service isn't in shutdown runlevel, so it stops before switching to shutdown runlevel. However, if signageos-clock lists it in it's dependencies, it will start again, which is unnecessary and also it takes a long time (about 1 minute in this case for some reason) which slows down the shutdown a lot. It works even without the dependency so I'm just removing it. The shutdown is much faster after that.