quick fix deal with already existing data partition on firstboot
there's some kind of bug that if the data partition already exists on first boot (for example after interrupted overwrite system process) it will be automatically mounted even though it shouldn't because it's not in fstab. However I don't have time right now to do thorough debugging so this is a good-enough quick fix. After next reboot, it should