implement independent firmware version
firmware version should reflect what version was flashed originally, since some things cannot be updated through app upgrade. This way it will be easier to debug problems remotely because we'll know which version the user originally flashed and we'll be able to determine if it has anything to do with the problems he might experience.