dont send signal to the server when automounting all usb devices during boot
sometimes the timing is so bad that the signal is sent exactly between the time when the node process started and when it actually binds the signal so when that happens, the node process is killed instead. We don't really need to notify it about a usb device being mounted during boot since it's too early anyway. Whatever needs to know about the usb devices, it didn't event have a chance to send a regular listStorageUnits() at that point.