implement custom smart cpu governor that prevents cpu from overheating
if we simply set ondemand governor, it will blast the cpu with highest frequency if there's high load for any extended period of time. However, running on high frequency for too long will overheat the device. This service will prevent that by implementing a "killswitch" that will set it back to powersave governor if the cpu gets too hot.